requirements.txt
google-generativeai>=0.9.0
python-pptx>=0.6.21
asyncio
pathlib 